backend/requirements.txt
addict==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
confection==1.3.3
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
elevenlabs==2.40.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro==0.9.4
language-tags==1.2.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
misaki==0.9.4
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.4.3
openai==2.30.0
openai-whisper==20250625
packaging==26.0
phonemizer-fork==3.3.2
preshed==3.0.13
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.33.0
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.1
segments==2.4.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
spacy==3.8.14
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
sympy==1.14.0
tavily-python==0.7.23
termcolor==3.3.0
thinc==8.3.13
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==1.0.0
websockets==16.0
wrapt==2.1.2
